pyproject.toml
[project]
name = "mmwave_radar_processing"
version = "0.1.0"
description = ""
authors = [
    {name = "David Hunt", email = "dmh89@duke.edu"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "numpy>=1.25,<2.0.0",
    "scikit-learn>=1.3.2,<2.0.0",
    "SciPy>=1.11.4,<2.0.0",
    "matplotlib>=3.8.2,<4.0.0",
    "tqdm>=4.66.1,<5.0.0",
    "ipywidgets>=8.1.1,<9.0.0",
    "jupyter>=1.0.0,<2.0.0",
    "pytest>=7.4.4,<8.0.0",
    "imageio[ffmpeg]>=2.33.1,<3.0.0",
    "addict>=2.4.0,<3.0.0",
    "pandas>=2.1.4,<3.0.0",
    "opencv-python>4.7,<4.9.0.80",
    "poetry-dotenv-plugin>=0.2.0,<0.3.0",
    "plotly>=6.0.1,<7.0.0",
    "open3d (>=0.19.0,<0.20.0)",
    "pyqtgraph (>=0.13.7,<0.14.0)",
    "pyqt6 (>=6.4.0,<7.0.0)",
    "imageio[ffmpeg] (>=2.37.2,<3.0.0)",
    "pyopengl (>=3.1.10,<4.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)",
]

[tool.poetry]
packages = [{include = "mmwave_radar_processing"}]

[tool.poetry.group.submodules]
optional = true

[tool.poetry.group.submodules.dependencies]
cpsl_datasets = {path = "./submodules/cpsl_datasets", develop = true}

[tool.poetry.group.dev.dependencies]
ipywidgets = ">=8.1.1,<9.0.0"
jupyter = ">=1.0.0,<2.0.0"

[tool.poetry.group.test.dependencies]
pytest = ">=7.4.4,<8.0.0"
bandit = ">=1.7,<2.0.0"
black = ">=22.12.0,<23.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
